Move more alloc tests out of runtests.c
