+++ expat/tests/handlers.h
@@ -363,6 +363,12 @@ extern int XMLCALL external_entity_alloc_set_encoding(XML_Parser
                                                       const XML_Char *systemId,
                                                       const XML_Char *publicId);
 
+extern int XMLCALL external_entity_reallocator(XML_Parser parser,
+                                               const XML_Char *context,
+                                               const XML_Char *base,
+                                               const XML_Char *systemId,
+                                               const XML_Char *publicId);
+
 /* Declaration handlers for entity declarations */
 extern void XMLCALL entity_suspending_decl_handler(void *userData,
                                                    const XML_Char *name,
